int -> double
Took 3 minutes

+++ app/lib/data/Data.dart
@@ -1,4 +1,5 @@
 import 'dart:convert';
+import 'dart:ffi';
 
 import 'package:flutter/material.dart';
 
@@ -114,7 +115,7 @@ class Reservation {
 
 class Invoice {
   String reservation;
-  int? price;
+  double? price;
   String? customer;
   DateTime? createdAt;
   DateTime? paidAt;
